+++ packages/koot/React/root.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Provider } from 'react-redux'
-import { Router } from 'react-router'
+import Router from 'react-router/lib/Router'
 
 import { StyleMapContext } from './styles'