+++ packages/koot/CHANGELOG.md
@@ -17,6 +17,14 @@
       - 这些 CSS 文件必须有一个名为 `.component` 或 `.[name]__component` 的 className
         - 该 className 会被更换为 hash 结果，如 `.a85c6k` 或 `.nav__bjj15a`
       - `props.className` 会传入到对应的组件，其值为与上述结果对应的 hash 后的 className
+- React SPA
+  - 确保
+- 更新依赖包
+  - minor
+    - `less` -> _3.9.0_
+  - patch
+    - `inquirer` -> _6.2.1_
+    - `portfinder` -> _1.0.20_
 
 ## 0.7.5
 **2018-12-03**